feat(dashboard-kyc-upload): update application flow for loans
