+++ components/LoanApplication/application/stepThree.tsx
@@ -4,7 +4,7 @@ import NextButton from "../steppedForm/nextButton";
 import ErrorMessage from "@/components/ui/error-message";
 import { Input } from "@/components/ui/input";
 import { useMultiStepForm } from "@/hooks/use-stepped-form";
-import { CombinedCheckoutSchema } from "@/validators/application-flow.validator";
+import { CombinedKycSchema } from "@/validators/application-flow.validator";
 import { useFormContext } from "react-hook-form";
 import { z } from "zod";
 import PrevButton from "../steppedForm/prevButton";
@@ -15,7 +15,7 @@ const Step3 = () => {
     // getValues,
     // setError,
     formState: { errors },
-  } = useFormContext<z.infer<typeof CombinedCheckoutSchema>>();
+  } = useFormContext<z.infer<typeof CombinedKycSchema>>();
 
   const { nextStep } = useMultiStepForm();
 
@@ -27,89 +27,42 @@ const Step3 = () => {
     <div className="pt-4 pb-8">
       <div className="grid sm:grid-cols-2 sm:gap-3">
         <div>
-          <label className="font-semibold">Job Title</label>
-          <Input {...register("jobTitle")} placeholder="e.g Accountant" />
-          <ErrorMessage message={errors.jobTitle?.message} />
+          <label className="font-semibold">First Name</label>
+          <Input {...register("kinFirstName")} placeholder="Simon" />
+          <ErrorMessage message={errors.kinFirstName?.message} />
         </div>
         <div>
-          <label className="font-semibold">Ministry</label>
+          <label className="font-semibold">Last Name</label>
+          <Input {...register("kinLastName")} placeholder="Mwansa" />
+          <ErrorMessage message={errors.kinLastName?.message} />
+        </div>
+        <div>
+          <label className="font-semibold">Phone Number</label>
           <Input
-            {...register("ministry")}
-            placeholder="e.g Ministry of Finance"
+            className="loan-number-input"
+            type="number"
+            inputMode="numeric"
+            {...register("kinPhoneNumber", {
+              valueAsNumber: true, // Convert to number automatically
+              validate: (value) =>
+                !isNaN(value) || "Please enter a valid number",
+            })}
+            placeholder="0978236744"
           />
-          <ErrorMessage message={errors.ministry?.message} />
+          <ErrorMessage message={errors.kinPhoneNumber?.message} />
         </div>
         <div>
-          <label className="font-semibold">Department</label>
-          <Input {...register("department")} placeholder="e.g Finance" />
-          <ErrorMessage message={errors.department?.message} />
+          <label className="font-semibold">Relationship</label>
+          <Input {...register("relationship")} placeholder="e.g Brother" />
+          <ErrorMessage message={errors.relationship?.message} />
         </div>
         <div>
-          <label className="font-semibold">Employee Number</label>
-          <Input {...register("employeeNumber")} placeholder="e.g 10336622" />
-          <ErrorMessage message={errors.employeeNumber?.message} />
-        </div>
-      </div>
-      {/* references */}
-      <div className="py-4">
-        <h2 className="text-center font-semibold text-purple-900">
-          References
-        </h2>
-        <h3 className="font-semibold py-3">Human Resource Manager</h3>
-        <div className="flex sm:flex-row flex-col sm:justify-between sm:flex-wrap md:flex-nowrap">
-          <div>
-            <label className="font-semibold">First Name</label>
-            <Input {...register("hrmFirstName")} placeholder="Simon" />
-            <ErrorMessage message={errors.hrmFirstName?.message} />
-          </div>
-          <div>
-            <label className="font-semibold">Last Name</label>
-            <Input {...register("hrmLastName")} placeholder="Mwansa" />
-            <ErrorMessage message={errors.hrmLastName?.message} />
-          </div>
-          <div>
-            <label className="font-semibold">Phone Number</label>
-            <Input
-              className="loan-number-input"
-              type="number"
-              inputMode="numeric"
-              {...register("hrmPhoneNumber", {
-                valueAsNumber: true, // Convert to number automatically
-                validate: (value) =>
-                  !isNaN(value) || "Please enter a valid number",
-              })}
-              placeholder="0978236744"
-            />
-            <ErrorMessage message={errors.hrmPhoneNumber?.message} />
-          </div>
-        </div>
-        <h3 className="font-semibold py-3">Supervisor</h3>
-        <div className="flex sm:flex-row flex-col sm:justify-between sm:flex-wrap md:flex-nowrap">
-          <div>
-            <label className="font-semibold">First Name</label>
-            <Input {...register("supervisorFirstName")} placeholder="Simon" />
-            <ErrorMessage message={errors.supervisorFirstName?.message} />
-          </div>
-          <div>
-            <label className="font-semibold">Last Name</label>
-            <Input {...register("supervisorLastName")} placeholder="Mwansa" />
-            <ErrorMessage message={errors.supervisorLastName?.message} />
-          </div>
-          <div>
-            <label className="font-semibold">Phone Number</label>
-            <Input
-              className="loan-number-input"
-              type="number"
-              inputMode="numeric"
-              {...register("supervisorPhoneNumber", {
-                valueAsNumber: true, // Convert to number automatically
-                validate: (value) =>
-                  !isNaN(value) || "Please enter a valid number",
-              })}
-              placeholder="0978236744"
-            />
-            <ErrorMessage message={errors.supervisorPhoneNumber?.message} />
-          </div>
+          <label className="font-semibold">Address</label>
+          <Input
+            {...register("kinAddress")}
+            placeholder="e.g Plot 451 Kabwata, Lusaka"
+          />
+          <ErrorMessage message={errors.kinAddress?.message} />
         </div>
       </div>
       <div className="w-full flex items-center justify-between">
