+++ components/layout/dashboard/ui/rightSidebar.tsx
@@ -3,7 +3,27 @@ import Link from "next/link";
 import styles from "@/styles/Dashboard.module.css";
 import { Pie } from "./progressBar";
 
-const RightSideBar = () => {
+interface RightSideBarType {
+  userID: string;
+  userName: string;
+  handleApplyButtonClick: () => void;
+}
+
+const RightSideBar = ({
+  userID,
+  userName,
+  handleApplyButtonClick,
+}: RightSideBarType) => {
+  const handleRepayLoan = () => {
+    console.log("paid");
+  };
+  const handleTrackLoan = () => {
+    console.log("track");
+  };
+  const handleWithdrawal = () => {
+    console.log("withdraw");
+  };
+
   return (
     <div className="hidden lg:flex flex-col gap-[10px] pt-[10px] pr-[10px]">
       <Link
@@ -14,14 +34,15 @@ const RightSideBar = () => {
       </Link>
       <div className="flex flex-col gap-3 pl-5 pr-3 py-8 bg-white rounded-md w-full">
         <div className="flex flex-col gap-2">
-          <p className="text-m">Lute Chongo</p>
-          <p className="text-m pb-2 text-yellow-400">User ID: MF00001</p>
+          <p className="text-m">{userName}</p>
+          <p className="text-m pb-2 text-yellow-400">User ID: {userID}</p>
           <div className={`${styles.gradDiv}`} />
         </div>
         <div className="flex flex-col items-start gap-5 w-[210px]">
           <div
             className={`${styles.main} flex w-full gap-6 items-center rounded-md py-3 px-4 cursor-pointer`}
             id="repay-loan"
+            onClick={handleRepayLoan}
           >
             <Image
               src={"/Icons/Dollar.svg"}
@@ -34,6 +55,7 @@ const RightSideBar = () => {
           <div
             className={`${styles.main} flex w-full gap-6 items-center rounded-md py-3 px-4 cursor-pointer`}
             id="track-loan"
+            onClick={handleTrackLoan}
           >
             <Image
               src={"/Icons/Hourglass.svg"}
@@ -43,7 +65,8 @@ const RightSideBar = () => {
             />
             <span className="text-m">Track Loan</span>
           </div>
-          <div
+          <Link
+            href={"/dashboard/support"}
             className={`${styles.main} flex gap-3 justify-between items-center rounded-md py-3 px-4 cursor-pointer`}
           >
             <Image
@@ -53,16 +76,18 @@ const RightSideBar = () => {
               alt="headsets icon"
             />
             <span className="text-m">Customer Support</span>
-          </div>
+          </Link>
           <button
             id="apply-for-loan"
             className="w-full py-3 text-center bg-purple-500 hover:bg-purple-primary rounded-md text-white"
+            onClick={handleApplyButtonClick}
           >
             Apply for Loan
           </button>
           <button
             id="withdraw-amount"
             className="py-3 w-full text-center border border-purple-400 bg-white hover:bg-yellow-300 rounded-md text-black"
+            onClick={handleWithdrawal}
           >
             Withdraw
           </button>
